refactor: refactor memory folders
