+++ src/memory/primitives/relocatable.ts
@@ -1,6 +1,3 @@
-import { Felt } from './felt';
-import { NotImplementedError } from './types';
-
 export class RelocatableError extends Error {}
 export class OffsetUnderflow extends RelocatableError {}
 export class SegmentError extends RelocatableError {}
