+++ src/memory/types.ts
